Functions are faster & more reliable than aliases
As noted in the fish documentation, the "alias" command exists for
backwards compatibility with POSIX shells and usually isn't the best
method for aliasing commands. There have also been performance issues
with the "alias" command, as noted in the fish issue tracker:
https://github.com/fish-shell/fish-shell/issues/486

In general, fish functions are more reliable and provide better
performance.