Support custom auto-activation files. Fixes #5.
This changes the default auto-activation filename from .vfenv to the
more standard .venv, while at the same time providing a setting that
can be used to customize this filename to any arbitrary value. For any
folk who already have .vfenv files, this setting can be used to provide
backwards compatibility.