Fix begin statements in compound if clauses.
5dcb1fe7057 introduced a number of begin statements that are missing a
semicolon. Although these work in fish 2.1.0 and earlier, later
development introduces a new parser that treats arguments to begin as an
error (in line with the documentation).

Bringing the begin statements into line with the documentation keeps
them working on all versions of fish.

(See https://github.com/fish-shell/fish-shell/issues/1248 for
discussion.)