Fix autocomplete help text when 'functions' prints comments.
Sometimes the 'functions' command outputs the function definition
location (file and line number) as a comment before the actual function
definition. When that happens, the help text would be set to the comment
instead of to the function description.

This greps for the first line starting with the word 'function' to find
the help text.