Document workon alias behavior w/ projects plugin
The purpose of this change is to help users understand how to get
workon` behaving similar to virtualenvwrapper's `workon` when the
`compat_aliases` and `projects` plugins are used together.

- The `compat_aliases` plug-in aliases `workon` to `vf activate`.
- The `project` plugin aliases `workon` to `vf workon`, but only
  when `VIRTUALFISH_COMPAT_ALIASES` is set when the plug-in is
  loaded.

So the order the two plugins are listed (i.e., loaded) matters,
and so does whether or not the VIRTUALFISH_COMPAT_ALIASES environment
variable is set.

If VIRTUALFISH_COMPAT_ALIASES is not set before the projects plugin is
loaded or the compat_aliases is loaded after the project plugin then
`workon` will activate the virtual environtment, but will not
automatically change to the  project directory.

See issue #137