Fix `workon` w/ Projects & Compat. Aliases enabled
Workon did not seem to behave as expected with both the Projects and
Compatibility Aliases plugins installed, even when loaded in the
proscribed order. Perhaps the first `workon` function definition "wins",
with subsequent definitions not overriding the first. This attempts to
solve that by first un-defining an existing definition, if defined.

This should also finally remove the last vestiges of the old
VIRTUALFISH_COMPAT_ALIASES setting, which had been removed from other
parts of the code a long time ago.