Env name no longer needs to be last arg to `vf new`
By looking at the provided arguments and making them to any Virtualenv
option flags take a value, we can safely determine which argument is the
virtual environment name without requiring that it be provided last.

When we pass the parsed arguments to Virtualenv, we now do so starting
with the environment name and then the option flags afterward.
Otherwise, using Virtualenv's --extra-search-dir option (which can take
multiple space-separated values) at the end of the invocation will cause
it to think the environment name is a path, causing an error.