Make Auto-Activation and Projects plugins play nice together (#175)
Activating both auto-activation and projects means that the
virtual environment of the project will be deactivated automatically
when leaving the project directory. This is accomplished by expanding
the Auto-Activation plugin to better detect projects in three steps:

1. Check if currently active virtualenv is connected to a project and
save its path

2. For cases where no activation file exists in the project directory,
additionally check if activation root is in project path to set
new_virtualenv_name

3. Finally, additionally check if any virtualenv is active while project
path doesn't match activation root. If so, auto-deactivate virtualenv