Remove temporary environments *safely* upon de-activation
If `vf tmp` failed due to a virtualenv error, activating another
environment would result in it being deleted upon de-activation due to
the naive manner in which the VF_TEMPORARY_ENV variable was set.

The renamed _VF_TEMPORARY_ENV variable is now set to the name of the
temporary environment, and upon de-activation the environment is only
deleted if the name stored in the _VF_TEMPORARY_ENV variable matches the
current environment name.

Adding a bit of extra safety, the `trash` command is used if present,
otherwise falling back to the more-destructive `rm -rf` command.