Use tool versions file to colorize `ls --details`
Previously, environment Python versions in the `vf ls --details` list
were only considered to be up-to-date (i.e., displayed in green) when a
given environment's Python version is greater or equal to the default
Python version returned by `__vfsupport_get_default_python()`. If you
have some environments on older but fully-supported Python versions,
those would show up in yellow even though they should be considered
up-to-date.

With the changes in this commit, given a `~/.tool-versions` file that
contains the following line:

    python 3.9.7 3.8.12 3.7.11 3.6.14

… any environment's Python version that matches one of the specified
versions will be considered up-to-date and displayed in green.